+++ analytics_on_fhir/settings.py
@@ -13,6 +13,9 @@ class StudyNames(Enum):
     STUDY_PROTOCOL_D = "study_protocol_d"
     STUDY_PROTOCOL_PCA1 = "study_protocol_pca1"
     STUDY_PROTOCOL_AML = "study_protocol_aml"
+    AML = "aml"
+    REACTO_COUNTS = "reacto_counts"
+    DQ = "dq"
     ALL = "all"
 
 
@@ -28,7 +31,19 @@ class SparkSettings:
 
 
 @ts.settings
+class FHIRServerSettings:
+    base_url: str | None = None
+    print_request_urls: bool = False
+    chunk_size: int = 100
+    page_count: int = 1000
+    num_processes: int = 6
+    user: str = ""
+    password: str = ts.secret(default="")
+
+
+@ts.settings
 class Settings:
+    fhir: FHIRServerSettings
     spark: SparkSettings
     aws_access_key_id: str = "admin"
     aws_secret_access_key: str = ts.secret(default="miniopass")